+++ HISTORY.md
@@ -6,6 +6,7 @@ https://github.com/josdejong/jsoneditor
 ## not yet published, version 6.1.0
 
 - Implemented menu options `sort` and `transform` for modes `code` and `text`.
+- Implemented new context menu item `extract`.
 
 
 ## 2019-06-12, version 6.0.0